+++ module/browse.php
@@ -84,6 +84,7 @@ if (isModEnabled('stock') && getDolGlobalInt('DOLICATALOG_SHOW_STOCK', 1)) {
 print '</div>';
 
 print '<div class="dolicatalog-browse-crumbs" id="dcb-breadcrumb"></div>';
+print '<div class="dolicatalog-facets" id="dcb-facets"></div>';
 print '<div class="dolicatalog-browse-body" id="dcb-results"></div>';
 print '<div class="dolicatalog-browse-pager" id="dcb-pager"></div>';
 
@@ -111,7 +112,7 @@ $labelKeys = array(
 	'DoliCatalogEmptyCategory', 'DoliCatalogError', 'DoliCatalogFavorites',
 	'DoliCatalogRecent', 'DoliCatalogOpenProduct', 'DoliCatalogAddFavorite',
 	'DoliCatalogRemoveFavorite', 'DoliCatalogPrevious', 'DoliCatalogNext',
-	'DoliCatalogBrowseEmpty',
+	'DoliCatalogBrowseEmpty', 'DoliCatalogRefineBy', 'DoliCatalogClearTags',
 );
 foreach ($labelKeys as $k) {
 	$config['labels'][$k] = $langs->transnoentities($k);